+++ iterator/SLT_single_string.h
@@ -23,10 +23,10 @@
  */
 typedef struct {
 	// Properties of the current right-maximal string W
-	unsigned long string_depth;
-	unsigned long bwt_start;
-	unsigned long revbwt_start;
-	unsigned long interval_size;
+	unsigned int string_depth;
+	unsigned int bwt_start;
+	unsigned int revbwt_start;
+	unsigned int interval_size;
 	unsigned char WL_char;  // ID of the label of the last Weiner link (i.e. of the first character of W). 0=#. 1..4: ACGT. Remark: this variable can never be 0.
 	
 	// Right extensions
@@ -36,10 +36,10 @@ typedef struct {
 	// Left extensions
 	unsigned char nleft_extensions;
 	unsigned char left_extension_bitmap;  // The i-th LSB is one iff the i-th character is a left extension. '#' is the 0-th LSB.
-	unsigned long left_ext_bwt_start[5];
+	unsigned int left_ext_bwt_start[5];
 	
 	// Frequency of every pair of left (rows) and right (columns) extension.
-	unsigned long left_right_extension_freqs[6][6];
+	unsigned int left_right_extension_freqs[6][6];
 } SLT_params_t;
